feature/juan/clean Feat: Update CharacterDetailActivity
