+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.detail.HeroDetailActivity"
+            android:name=".ui.detail.CharacterDetailActivity"
             android:parentActivityName=".ui.main.MainActivity"
             android:theme="@style/Theme.ArquitectoMarvel.NoActionBar" />
     </application>
