+++ app/src/main/res/layout/activity_character_detail.xml
@@ -20,8 +20,8 @@
             android:fitsSystemWindows="true"
             app:contentScrim="?attr/colorPrimary"
             app:layout_scrollFlags="scroll|exitUntilCollapsed"
-            app:title="@tools:sample/lorem/random"
-            app:toolbarId="@+id/toolbar">
+            app:toolbarId="@+id/toolbar"
+            tools:title="@tools:sample/lorem/random">
 
             <androidx.appcompat.widget.AppCompatImageView
                 android:id="@+id/header_hero_image"
@@ -30,15 +30,15 @@
                 android:contentDescription="@string/hero_image"
                 android:fitsSystemWindows="true"
                 android:scaleType="centerCrop"
-                android:src="@mipmap/ic_launcher"
-                app:layout_collapseMode="parallax" />
+                android:transitionName="@string/hero_image"
+                app:layout_collapseMode="parallax"
+                tools:src="@mipmap/ic_launcher" />
 
             <ImageView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:contentDescription="@string/hero_image"
-                android:src="@drawable/black_background"
-                android:transitionName="@string/hero_image" />
+                android:src="@drawable/black_background" />
 
             <androidx.appcompat.widget.Toolbar
                 android:id="@+id/toolbar"
@@ -46,7 +46,7 @@
                 android:layout_height="?attr/actionBarSize"
                 app:layout_collapseMode="pin"
                 app:popupTheme="@style/Theme.ArquitectoMarvel.PopupOverlay"
-                app:title="@tools:sample/lorem/random" />
+                tools:title="@tools:sample/lorem/random" />
 
         </com.google.android.material.appbar.CollapsingToolbarLayout>
     </com.google.android.material.appbar.AppBarLayout>
@@ -63,6 +63,6 @@
 
     <include
         android:id="@+id/content_hero_detail"
-        layout="@layout/content_hero_detail" />
+        layout="@layout/content_character_detail" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
