+++ app/src/main/res/layout/content_character_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:showIn="@layout/activity_hero_detail">
+    tools:showIn="@layout/activity_character_detail">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -71,9 +71,5 @@
                 tools:listitem="@layout/comic_item" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
-
-
     </LinearLayout>
-
-
 </androidx.core.widget.NestedScrollView>
